:truck: Move and rename SnoozeTask use case
Snooze Task was renamed to Snooze Alarm and moved to a package that
better fits it
