+++ data/domain/src/main/java/com/escodro/domain/di/DomainModule.kt
@@ -6,6 +6,7 @@ import com.escodro.domain.usecase.alarm.RescheduleFutureAlarms
 import com.escodro.domain.usecase.alarm.ScheduleAlarm
 import com.escodro.domain.usecase.alarm.ScheduleNextAlarm
 import com.escodro.domain.usecase.alarm.ShowAlarm
+import com.escodro.domain.usecase.alarm.SnoozeAlarm
 import com.escodro.domain.usecase.alarm.UpdateTaskAsRepeating
 import com.escodro.domain.usecase.category.DeleteCategory
 import com.escodro.domain.usecase.category.LoadAllCategories
@@ -15,7 +16,6 @@ import com.escodro.domain.usecase.task.AddTask
 import com.escodro.domain.usecase.task.CompleteTask
 import com.escodro.domain.usecase.task.DeleteTask
 import com.escodro.domain.usecase.task.GetTask
-import com.escodro.domain.usecase.task.SnoozeTask
 import com.escodro.domain.usecase.task.UncompleteTask
 import com.escodro.domain.usecase.task.UpdateTask
 import com.escodro.domain.usecase.task.UpdateTaskStatus
@@ -37,7 +37,6 @@ val domainModule = module {
     factory { UpdateTaskStatus(get(), get(), get()) }
     factory { DeleteTask(get(), get()) }
     factory { GetTask(get()) }
-    factory { SnoozeTask(get(), get(), get()) }
     factory { UpdateTask(get()) }
 
     // Category Use Cases
@@ -56,8 +55,9 @@ val domainModule = module {
     factory { RescheduleFutureAlarms(get(), get(), get()) }
     factory { ScheduleAlarm(get(), get()) }
     factory { ScheduleNextAlarm(get(), get(), get()) }
-    factory { UpdateTaskAsRepeating(get()) }
     factory { ShowAlarm(get(), get(), get()) }
+    factory { SnoozeAlarm(get(), get(), get()) }
+    factory { UpdateTaskAsRepeating(get()) }
 
     // Tracker Use Cases
     factory { LoadCompletedTasksByPeriod(get()) }
